[MS] Renaming WorkspaceTagRole to WorkspaceRoleTag & TagProfile to UserProfileTag
