+++ client/tests/component/specs/testTagProfile.spec.ts
@@ -2,14 +2,14 @@
 
 import { getDefaultProvideConfig } from '@tests/component/support/mocks';
 
-import TagProfile from '@/components/users/TagProfile.vue';
+import UserProfileTag from '@/components/users/UserProfileTag.vue';
 import { UserProfile } from '@/parsec';
 import { mount } from '@vue/test-utils';
 import { describe, expect, it } from 'vitest';
 
 describe('User Avatar', () => {
   it('Display avatar for user', () => {
-    const wrapper = mount(TagProfile, {
+    const wrapper = mount(UserProfileTag, {
       props: {
         profile: UserProfile.Outsider,
       },
