+++ client/tests/component/specs/testWorkspaceTagRole.spec.ts
@@ -2,14 +2,14 @@
 
 import { getDefaultProvideConfig } from '@tests/component/support/mocks';
 
-import WorkspaceTagRole from '@/components/workspaces/WorkspaceTagRole.vue';
+import WorkspaceRoleTag from '@/components/workspaces/WorkspaceRoleTag.vue';
 import { WorkspaceRole } from '@/parsec';
 import { mount } from '@vue/test-utils';
 import { describe, expect, it } from 'vitest';
 
 describe('User Avatar', () => {
   it('Display avatar for user', () => {
-    const wrapper = mount(WorkspaceTagRole, {
+    const wrapper = mount(WorkspaceRoleTag, {
       props: {
         role: WorkspaceRole.Manager,
       },
